Replace acquiring csrf token from cookies with page payload
As a SPA we have to include the CSRF token in a X-CSRFToken header
instead of a hidden form field.  However the previous code acquired
the token value from the cookie,  which fails if the cookie is marked
httponly.  Instead we add the token to the app-state via the page
payload, and use it from there instead.