DUMA: extend user-defined test to include category
We now define a user-added term, for the purposes of DUMA integration,
as one with the isUserDefined flag set to true, *and* a
userAddedCategory field.  The latter field is so that the integration
can know what an arbitrary user-added term is (parameter, platform,
unit,...) and also means that terms not of interest to DUMA can still
have isUserDefined set but not be returned.  Bounding-boxes are a
canonical example here.