User-added Modals: parameters (WIP): all user-editable fields mandatory
This is a special-case, because there is a nested units field, which
can also be user-added.  If we added required-field logic to all of
those fields, then when we select a pre-existing unit the user-added
form will be empty and thus errors will remain.

Note that this case may not make sense anyway, because it implies that
the unit is tied to the parameter, which is not reflected at all in
the XML.