About5+6: Bugfix: xpath can't index on target node
An index on the *target* node, eg additionalDocumentation[1], is
flawed because this xpath -- with the data/element index -- used to
select the next element in the array to export into.  The logic is
basically node.xpath()[index], but the use of an index in the xpath
expression means it will only return one element and thus any array
size more than 1 will fail.