Who: WIP: updating organisation defs with ES/API field names
The bigger issue remaining is the "." top-level xpath.  There may be a
way around this, but way I've been structuring arrays so far is for
the array element and the items-object to both have the same xpath,
which doesn't seem feasible as currently-structured.  May just copy
and paste, and stop worrying about duplication.