Add rule to support person-or-organisation case
This case (which occurs in citedResponsibleParty and pointOfContact)
allows selecting either an individual or an organisation, and due to
the way the schema is set up using simple field-level "required field"
rules would mean the non-selected case would cause errors to be
displayed.

Instead, we hard-code the logic based on the party-type and be done
with it.