DUMA: hack to preserve the submitted=false testing flag
We introduced this flag to return documents that haven't been
submitted to facilitate testing, but as a simple filter it wasn't
preserved in subsequent hyperlinks. Hack it so if we are serialising a
document that hasn't been submitted, assume this flag was present and
append it so it is preserved all the way through.