Who7: Implement post-processing cleanup for orcid/uri case
A contact can have both an orcid, and a uri (from the controlled vocab
picker).  We want to export the orcid in preference, and fall back to
the URI; this is complicated by the fact that they have slightly
different export paths.  So, we include both options in the
template (ie, the template isn't valid before export) and export both,
to be cleaned up at the end.