Feature: experimental: "append" handling for array export
Allow simply using the last matching element rather than the specified
element-index (it is up to the "caller" to ensure that this is
correct, ie that the template element is inserted in the correct
location). This allows us to handle cases where the xpath requires a
position argument, which is incompatible with standard export.