Template updates: process steps changes
These are a bit further-reaching: we used to have separate
resourceLineage sections for the creation and process steps, but
they're all under the one roof.  Additionally, this means that we
don't have a summary/title/description for the process steps, so those
fields are removed from the mapper, and also from the UI.

Most of that is in the form of commenting-out, which should probably
be cleaned up later when we're sure.