Health-check version: use pre-existing GIT_VERSION
Belatedly I notice that the docker image is build with a GIT_VERSION
environment variable defined, which we could use.  I don't mind the
timestamp as well, but for now I'll leave the current operatoin in
place and if necessary we can remove the static-file creation from the
pipeline, and this will fall back to using the GIT_VERSION variable.

(Note that GIT_VERSION will be a lot less fresh locally, because we
rarely need to rebuild the image)