SD-73: DUMA API still needs to be strict about statuses returned
By default we return only submitted records, but for ease of testing
we support a query param to override this.  The implementation of this
was too lax, so it also returned archived, etc (no filtering on status
at all).  Change this so the default is submitted-only, and the query
override only adds draft docs to the mix.