SD-81: export of position
Two things treading on each others toes: I'd added the "position"
entry to the schema (from the ES index), but with a slightly incorrect
xpath. However also there was a "positionName" schema key with an
identical xpath (I'm guessing from before the index had stabilised),
which was clearing the value from position because it itself had no
data.