Bugfix: append handler: delete all template nodes
There's complications here, if that function is ever re-used -- the
problem with appending and templates is you assume that there's always
a template node, and that it can be deleted. If you ever use
traditional export *then* append you can't delete the template node,
but this is also unworkable since you may have had *nothing* to export
in the previous section.