SD-173: ignore DUMA key errors
This seemed prudent at the time, to be verbose if a document it didn't
know how to handle was encountered, but despite mitigation steps taken
during migration the likelihood that there'll be more, but harmless,
seems to outweigh the initial benefits.

So, we'll just warn and then continue while ignoring any path that
couldn't be processed.  If we hear about entries that aren't showing
up in the DUMA API, we can revisit.