SD-192: Use escape-hatch handler for additional constraints
This was a complicated situation because there are an arbitrary number
of static nodes (that must be read and displayed read-only), meaning
the xpath uses "position()<last()" and other tricky elements.  In
conjunction with needing to delete the template element, we were
inadvertently deleting the static nodes as well when there are no
additional-constraints.

It still feels that a "pure" solution should be feasible, but given
time constraints we'll just utilise the escape-hatch.