IWA-897: handle all cases of distributionInfo export
Ie, only data format present, only data sources present, both present,
both absent.  Multiple overlapping concerns made this quite tricky to
get right, including a refinement of the remove_element handler (which
is supposed to remove an element at an xpath, but only if the data
path isn't present... however this can result in situations where a
null value is stored so the path exists but the value should be
considered missing).  (I'm nervous that I'm misunderstanding this, or
that there are subtle bugs that this change may expose!)