Bugfix: cold start migrations attempt to find non-existent superuser
There's actually a couple of issues here, and the solution is 100%
effective, but it's close enough.  The root problem is that migrations
fail from a cold start (empty database), and repeated restarts don't
help.

The main problem there is a migration that runs code, and attempts to
find a superuser in order to save records requiring an associated user
id.  Since we haven't set up a user yet, this fails.

Subsequent migrations also fail, even though the django_entrypoint
script sets up a superuser, because the "find an admin" routine does
so by looking for a *staff* user, which a superuser is not. The fix is
to look for a superuser not a staff (we could have ensured that the
first superuser also had the staff flag set), since that seems in line
with the intent.

Fix applied in the TERN branch as well, even though that one isn't
used any more.